refactored directory structure
